Migrate database from SQLite to Postgres
Swaps better-sqlite3 for postgres.js, rewrites the Drizzle schema for
pg-core (native text[] columns for tags/photoUrls, jsonb for nested
ingredients/steps, timestamp columns), and regenerates migration
history for the new dialect. Adds docker-compose for local Postgres
and updates the Render blueprint to provision a managed database
instead of storing SQLite on a disk.

Also fixes two latent bugs the dialect change surfaced: Postgres
returns avg()/count() as strings rather than numbers, which would
have broken RecipeRating's average.toFixed(1) and collection recipe
counts — both now cast to float8/int in the query.

+++ src/collections/collections.server.ts
@@ -8,7 +8,7 @@ export async function findCollectionsByOwner(ownerId: string) {
       id: collections.id,
       name: collections.name,
       createdAt: collections.createdAt,
-      recipeCount: sql<number>`count(${collectionRecipes.recipeId})`,
+      recipeCount: sql<number>`count(${collectionRecipes.recipeId})::int`,
     })
     .from(collections)
     .leftJoin(collectionRecipes, eq(collectionRecipes.collectionId, collections.id))
